Use Poetry for dependency management and packaging (SVRENG-252) (#64)
* Use Poetry for dependency management and packaging

Alters the package to use Poetry for dependency management and packaging.
Reasoning for this includes, but is not limited to, the fact that it is
not really advisable to develop in an environment where the `make develop`
incantation makes changes to system or user Python packages, which can
quickly lead to dependency conflicts. (I've been hit by this a few times.)

Resolves #63.

* Update GitHub Actions workflow configuration to use Poetry

Additionally alters the Makefile to force commands--when applicable--to
run within the Poetry venv via `poetry run`.