boardwalkd (enh/security): Enable click generation of envvars for options
Specifies context settings for the Click cli.command() for `serve` in
boardwalkd, so that any click.options have environment variables automatically
created and checked for values, using the `BOARDWALKD` prefix.

Tagged as security-related as the key driving factor behind this change was to
enable use of the `--slack-webhook-url` and `--slack-error-webhook-url` from
envvars, because the webhooks URLs are considered secret, and consequently they
should not be passed on the command line where they can wind up in the shell history.

Resolves #69.
Resolves Backblaze SVRENG-270.