boardwalk (feat): display --limit from CLI
It is currently possible to provide a `--limit` when calling the 'boardwalk'
command, to further restrict and sub-select the host pattern present within a
Workspace's configuration. However, prior to this commit, the contents, if any,
of the limit were not displayed on the boardwalkd web UI.

Thus, the `worker_limit` is added to the WorkspaceDetails model, thus allowing
for the `--limit` (if any) to be sent to boardwalkd upon workflow execution.

As we use the limit as directly defined from the worker--and as it is possible
to _not_ define a limit, we display 'unknown' on the UI if the limit is not
defined. That could be caused by the aforementioned undefined limit--thus being
equivalent to an `all` limit--or because the boardwalk client was not updated,
and thus does not send the limit data in the WorkspaceDetails sent to the server
upon initiation of a workflow.

Resolves Issue #56.