Enable tab completion of commands/workspace names (#81)
The `click` framework provides tab completion for more or less free, however
it does need to be enabled. This PR adds the requirements to set up tab
completion to the README.md, as well as adds completion support for
Workspace names to the `boardwalk workspace use` command. Prior to this,
knowing the full workspace name would be required, thus requiring the user
to look in the Boardwalkfile, or run `boardwalk workspace list` beforehand.

Resolves internal ticket SVRENG-493.