documentation: Use Sphinx for building HTML documentation
Implements Sphinx to have some HTML-based documentation, so referencing documentation that is
not directly in Python source code files is possible. While this branch is not intended to
fully document all available functions, this _does_ provide the necessary framework which
allows future documentation of the code -- within the code itself by way of docstrings -- to
auto-generate updated documentation.

Additionally, while not strictly needed -- and I'll probably look into making this a
change that's available in boardwalk/boardwalkd themselves by default -- I built some
tooling to make pretty HTML output of the `--help` documentation for commands using
rich-click. Right now this isn't baked into the existing documentation build process,
though I _might_ look into seeing if it can be in the future. (I made an attempt, at
least, but the Bash script which does this was the 'good enough' compromise, since
it felt like I was fighting with Python for an excessive length of time to get it to
do what I wanted it to do.

Changes that would be needed on the main Backblaze/boardwalk repository are the usual
ones when enabling GitHub Pages; for my fork, this was -- after the first run of the
workflow to build and populate the `gh-pages` branch -- set the Repository's GitHub
Pages settings to deploy from a branch, using the `gh-pages` branch, and from the root.

Aside from that, the Sphinx documentation will auto-publish on push to the main branch,
or on a version tag being pushed/created.