Fix incorrect -- for Python 3.11 -- syntax for f-string in docs/conf.py.
This was an issue caused by the fact that my local development environment is running,
surprise surprise, Python 3.12, and the Actions workflows are running 3.11. PEP 701,
implemented in Python 3.12, makes it so the following pattern would function as expected.

```python
release = f"{lib_version("boardwalk")}"
```

https://docs.python.org/3/whatsnew/3.12.html#pep-701-syntactic-formalization-of-f-strings

As a note discovered while researching this issue, Python 3.13 is now released; as such,
eventually -- not in this PR, which is effectively just documentation -- the minimum
Python version will be shifted to 3.12.