documentation: Use Sphinx for building HTML documentation (#87)
* documentation: Use Sphinx for building HTML documentation

Backwards compatible; the only (minor) alteration to `boardwalk` -- which is the only reason this is getting a
version increment -- is invocations of `boardwalk` now only emit the current loglevel -- that is,
`Log level is {loglevel}`, at verbosity 1 or higher (e.g., `boardwalk -v ...`).

Implements Sphinx to have some HTML-based documentation, so referencing documentation that is
not directly in Python source code files is possible. While is not intended to fully document all available
functions with this change, this _does_ provide the necessary framework which
allows future documentation of the code -- within the code itself by way of docstrings -- to
auto-generate updated documentation.

This can be accomplished by running `make docs` from the root of the repository, or in the alternative,
running `make html` from the `docs/` directory (if in a `poetry shell` session). This directs Sphinx to
build the HTML documentation, which when complete is available at `docs/build/html/index.html`.