boardwalk (bug): Ensure workflow is set as succeeded false when beginning a workflow
This commit adds a test to ensure that when reusing a workspace -- for example, updating
the target version to upgrade an operating system to -- the remote state file is updated
to reflect the fact that the Workspace's Workflow has not yet successfully completed,
prior to said workflow actually beginning.

This ensures that if the reused Workspace's workflow completed successfully, a failure
within the current workflow won't requite manual user intervention to reset the remote
state file's workflow success attribute to false. This allows the workflow to be resumed,
as prior to this if a host met preconditions, and as a result of the workflow they
altered to a state where they are no longer met -- e.g., an OS upgrade -- the workflow
would refuse to allow resumption of the failed workflow, because -- to Boardwalk -- the
host already succeeded.

The code change will be added in the next commit.